+++ x/evmutil/types/events.go
@@ -8,6 +8,8 @@ const (
 	EventTypeConvertERC20ToCoin = "convert_evm_erc20_to_coin"
 	EventTypeConvertCoinToERC20 = "convert_evm_erc20_from_coin"
 
+	EventTypeConvertCosmosCoinToERC20 = "convert_cosmos_coin_to_erc20"
+
 	// Event Attributes - Common
 	AttributeKeyReceiver = "receiver"
 	AttributeKeyAmount   = "amount"